+++ src/Resources/contao/config/config.php
@@ -115,6 +115,7 @@ $GLOBALS['FE_MOD']['fernschachverwaltung'] = array
 (
 	'fernschachverwaltung_meldeformular'        => 'Schachbulle\ContaoFernschachBundle\Modules\Meldeformular_Spieler',
 	'fernschachverwaltung_meldeformular_team'   => 'Schachbulle\ContaoFernschachBundle\Modules\Meldeformular_Mannschaft',
+	'fernschachverwaltung_titel'                => 'Schachbulle\ContaoFernschachBundle\Modules\Titel',
 	'fernschachverwaltung_titelnormen'          => 'Schachbulle\ContaoFernschachBundle\Modules\TitelNormen',
 	'fernschachverwaltung_titelnormen_liste'    => 'Schachbulle\ContaoFernschachBundle\Modules\TitelNormenLast',
 	'fernschachverwaltung_kontoauszug'          => 'Schachbulle\ContaoFernschachBundle\Modules\Kontoauszug',