+++ src/Resources/contao/dca/tl_fernschach_spieler.php
@@ -208,7 +208,7 @@ $GLOBALS['TL_DCA']['tl_fernschach_spieler'] = array
 	'palettes' => array
 	(
 		'__selector__'                => array('death', 'honor_25', 'honor_40', 'honor_50', 'honor_60', 'honor_70', 'honor_president', 'honor_member', 'sepaBeitrag', 'sepaNenngeld', 'isDeletion', 'membercheck'),
-		'default'                     => '{archived_legend:hide},archived;{assign_legend:hide},memberAssign;{person_legend},nachname,vorname,titel,anrede,briefanrede,status;{live_legend},birthday,birthplace,sex,death;{adresse_legend:hide},plz,ort,bundesland,strasse,adresszusatz;{adresse2_legend:hide},plz2,ort2,bundesland2,strasse2,adresszusatz2;{telefon_legend:hide},telefon1,telefon2;{telefax_legend:hide},telefax1,telefax2;{email_legend:hide},email1,email2;{memberships_legend},memberId,memberInternationalId,isDeletion,patron,memberships,verein,membercheck;{alternativ_legend:hide},gastNummer,servertesterNummer,fremdspielerNummer;{zuzug_legend:hide},zuzug;{turnier_legend:hide},spielberechtigungen,klassenberechtigung,turnierAnmeldungenBewerbungen;{iccf_legend:hide},titelinfo;{normen_legend},normen;{honors_legend},honor_25,honor_40,honor_50,honor_60,honor_70,honor_president,honor_member;{bank_legend:hide},inhaber,iban,bic;{beitrag_legend},beitragsschulden,checkBeitrag,contribution_paid,beitrag2026;{sepaBeitrag_legend:hide},sepaBeitrag;{sepaNenngeld_legend:hide},sepaNenngeld;{download_legend},downloads;{info_legend:hide},info_beitritt,info;{publish_legend},published'
+		'default'                     => '{archived_legend:hide},archived;{assign_legend:hide},memberAssign;{person_legend},nachname,vorname,titel,anrede,briefanrede,status;{live_legend},birthday,birthplace,sex,death;{adresse_legend:hide},plz,ort,bundesland,strasse,adresszusatz;{adresse2_legend:hide},plz2,ort2,bundesland2,strasse2,adresszusatz2;{telefon_legend:hide},telefon1,telefon2;{telefax_legend:hide},telefax1,telefax2;{email_legend:hide},email1,email2;{memberships_legend},memberId,memberInternationalId,isDeletion,patron,memberships,verein,membercheck;{alternativ_legend:hide},gastNummer,servertesterNummer,fremdspielerNummer;{zuzug_legend:hide},zuzug;{turnier_legend:hide},spielberechtigungen,klassenberechtigung,turnierAnmeldungenBewerbungen;{qualifikationen_legend:hide},qualifikationen;{iccf_legend:hide},titelinfo;{normen_legend},normen;{honors_legend},honor_25,honor_40,honor_50,honor_60,honor_70,honor_president,honor_member;{bank_legend:hide},inhaber,iban,bic;{beitrag_legend},beitragsschulden,checkBeitrag,contribution_paid,beitrag2026;{sepaBeitrag_legend:hide},sepaBeitrag;{sepaNenngeld_legend:hide},sepaNenngeld;{download_legend},downloads;{info_legend:hide},info_beitritt,info;{publish_legend},published'
 	),
 
 	// Subpalettes
@@ -1068,6 +1068,117 @@ $GLOBALS['TL_DCA']['tl_fernschach_spieler'] = array
 				'tl_class'            => 'w50',
 			),
 			'sql'                     => "varchar(20) NOT NULL default ''"
+		),
+		'qualifikationen' => array
+		(
+			'label'                   => &$GLOBALS['TL_LANG']['tl_fernschach_spieler']['qualifikationen'],
+			'exclude'                 => true,
+			'inputType'               => 'multiColumnWizard',
+			'eval'                    => array
+			(
+				'tl_class'            => 'clr',
+				'buttonPos'           => 'top',
+				'columnFields'        => array
+				(
+					'fuer' => array
+					(
+						'label'                   => &$GLOBALS['TL_LANG']['tl_fernschach_spieler']['qualifikation_fuer'],
+						'exclude'                 => true,
+						'search'                  => false,
+						'sorting'                 => true,
+						'flag'                    => 12,
+						'inputType'               => 'text',
+						'eval'                  => array
+						(
+							'style'             => 'width: 200px',
+						),
+					),
+					'im_turnier' => array
+					(
+						'label'                   => &$GLOBALS['TL_LANG']['tl_fernschach_spieler']['qualifikation_im_turnier'],
+						'exclude'                 => true,
+						'search'                  => false,
+						'sorting'                 => true,
+						'flag'                    => 12,
+						'inputType'               => 'text',
+						'eval'                  => array
+						(
+							'style'             => 'width: 200px',
+						),
+					),
+					'vom' => array
+					(
+						'label'                 => &$GLOBALS['TL_LANG']['tl_fernschach_spieler']['qualifikation_vom'],
+						'exclude'               => true,
+						'inputType'             => 'text',
+						'eval'                    => array
+						(
+							'maxlength'           => 10,
+							'style'               => 'width: 100px',
+							'rgxp'                => 'alnum'
+						),
+						'load_callback'           => array
+						(
+							array('\Schachbulle\ContaoHelperBundle\Classes\Helper', 'getDate')
+						),
+						'save_callback' => array
+						(
+							array('\Schachbulle\ContaoHelperBundle\Classes\Helper', 'putDate')
+						),
+					),
+					'gueltig_bis' => array
+					(
+						'label'                 => &$GLOBALS['TL_LANG']['tl_fernschach_spieler']['qualifikation_gueltig_bis'],
+						'exclude'               => true,
+						'inputType'             => 'text',
+						'eval'                    => array
+						(
+							'maxlength'           => 10,
+							'style'               => 'width: 100px',
+							'rgxp'                => 'alnum'
+						),
+						'load_callback'           => array
+						(
+							array('\Schachbulle\ContaoHelperBundle\Classes\Helper', 'getDate')
+						),
+						'save_callback' => array
+						(
+							array('\Schachbulle\ContaoHelperBundle\Classes\Helper', 'putDate')
+						),
+					),
+					'genutzt_fuer' => array
+					(
+						'label'                 => &$GLOBALS['TL_LANG']['tl_fernschach_spieler']['qualifikation_genutzt_fuer'],
+						'exclude'               => true,
+						'inputType'             => 'text',
+						'eval'                  => array
+						(
+							'style'             => 'width: 200px',
+						),
+					),
+					'angemeldet_am' => array
+					(
+						'label'                 => &$GLOBALS['TL_LANG']['tl_fernschach_spieler']['qualifikation_angemeldet_am'],
+						'exclude'               => true,
+						'inputType'             => 'text',
+						'eval'                    => array
+						(
+							'maxlength'           => 10,
+							'style'               => 'width: 100px',
+							'rgxp'                => 'alnum'
+						),
+						'load_callback'           => array
+						(
+							array('\Schachbulle\ContaoHelperBundle\Classes\Helper', 'getDate')
+						),
+						'save_callback' => array
+						(
+							array('\Schachbulle\ContaoHelperBundle\Classes\Helper', 'putDate')
+						),
+					),
+				)
+			),
+			'sql'                   => "blob NULL"
 		),
 		'titelinfo' => array
 		(