+++ src/Resources/contao/languages/de/modules.php
@@ -30,6 +30,7 @@ $GLOBALS['TL_LANG']['FMD']['fernschachverwaltung_meldeformular_team'] = array('M
 $GLOBALS['TL_LANG']['FMD']['fernschachverwaltung_titelnormen'] = array('Titel und Normen ausgeben','Ausgabe der Titel und Normen aus der Fernschach-Verwaltung');
 $GLOBALS['TL_LANG']['FMD']['fernschachverwaltung_titelnormen_liste'] = array('Glückwunschliste Titel und Normen','Ausgabe der Titel und Normen aus der Fernschach-Verwaltung');
 $GLOBALS['TL_LANG']['FMD']['fernschachverwaltung_kontoauszug'] = array('Kontoauszug BdF-Mitglied', 'Kontoauszug des angemeldeten Mitglieds anzeigen.');
+$GLOBALS['TL_LANG']['FMD']['fernschachverwaltung_titel'] = array('Liste der Titelträger', 'Liste der Titelträger anzeigen.');
 
 $GLOBALS['TL_LANG']['MOD']['iccf_import_headline'] = array('Wertungszahlen aus CSV-Dateien importieren', 'Wertungszahlen aus CSV-Dateien importieren');
 $GLOBALS['TL_LANG']['MOD']['iccf_import_file'] = array('CSV-Datei', 'CSV-Datei auswählen');